requirements.txt
matplotlib==3.8.4
numpy==2.0.0
protobuf==5.27.1
pytorch_lightning==2.2.1
PyYAML==6.0.1
ray==2.9.3
scikit_learn==1.3.2
setproctitle==1.3.3
sortedcontainers==2.4.0
tensorflow==2.16.1
torch==2.1.2
torch_cluster==1.6.3+pt21cu121
torch_geometric==2.5.3
tqdm==4.66.2